test: split zaptec_constants into tests/zaptec/conftest.py
sveinse asked for setup specific to the future-standalone API client
(#414 review) to live under tests/zaptec/, separate from the HA
integration test setup. zaptec_username/zaptec_password stay in the
root conftest since tests/test_diagnostics.py (outside tests/zaptec,
not yet converted to the new pattern) still needs them.
